feat: merge adrenaserver into adrenaline
this should make it less work to maintain both, simpler for users and admins, and better for modpacks that are intended to be installed on both the server and client.

this utilizes mrpack's side feature so that only client/both-side mods will be installed in launchers and only server/both-side mods will be installed on the server.

i will be updating adrenaline's wiki pages and description to include a notice about this and instructions about how to install it on the server later :)